scripts/phylogenetic_analysis.py
from Bio import Phylo


def align_sequences():
    pass


def build_phylogenetic_tree(alignments):
    # Build a phylogenetic tree from sequence alignments
    tree = Phylo.construct(alignments, method="neighbor-joining")
    return tree


def visualize_phylogenetic_tree(tree):
    # Visualize the phylogenetic tree
    Phylo.draw(tree)


# Example usage
if __name__ == "__main__":
    sequence_alignments = align_sequences()  # Replace with actual alignment function
    phylogenetic_tree = build_phylogenetic_tree(sequence_alignments)
    visualize_phylogenetic_tree(phylogenetic_tree)
