scripts/requirements.txt
biopython==1.82
matplotlib==3.8.2